Don't know why I can't add to movie list
