+++ Movie List/Info.plist
