+++ Movie List/Resources/Assets.xcassets/Clapboard.imageset/1045-clapboard@3x.png
