+++ Movie List/Resources/Assets.xcassets/Clapboard.imageset/Contents.json
