+++ Movie List/Resources/Assets.xcassets/Plus.imageset/Contents.json
