+++ Movie List/Resources/LambdaLogoAssets.xcassets/Contents.json
