+++ Movie List/Resources/LambdaLogoAssets.xcassets/Lambda_Logo_Full.imageset/Contents.json
