+++ Movie List/Resources/LambdaLogoAssets.xcassets/Lambda_Logo_Full.imageset/Lambda_Logo_Full-1.pdf
