+++ Movie List/View Controllers/MovieListTableViewController.swift
@@ -1,0 +1,57 @@
+//
+//  MovieListTableViewController.swift
+//  Movie List
+//
+//  Created by Rob Vance on 2/9/20.
+//  Copyright © 2020 Lambda School. All rights reserved.
+//
+
+import UIKit
+
+class MovieListTableViewController: UIViewController {
+
+//Mark: IBOutlet
+    @IBOutlet weak var movieTableView: UITableView!
+    
+    var movies: [Movie] = []
+    
+    
+    override func viewDidLoad() {
+        super.viewDidLoad()
+    }
+    
+    
+    override func prepare(for segue: UIStoryboardSegue, sender: Any?) {
+        if segue.identifier == "AddMovieSegue" {
+            if let addMovieVC = segue.destination as? AddMovieViewController {
+                addMovieVC.delegate = self
+            }
+        }
+    }
+}
+
+extension MovieListTableViewController: UITableViewDataSource {
+    func tableView(_ tableView: UITableView, numberOfRowsInSection section: Int) -> Int {
+        return movies.count
+
+    }
+    
+    func tableView(_ tableView: UITableView, cellForRowAt indexPath: IndexPath) -> UITableViewCell {
+        guard let cell = movieTableView.dequeueReusableCell(withIdentifier: "MovieTitleCell", for: indexPath) as? MovieTableViewCell else { return UITableViewCell() }
+        
+        let movie = movies[indexPath.row]
+        cell.movie = movie
+        return cell
+    }
+   
+}
+
+extension MovieListTableViewController: AddMovieDelegate {
+    func movieWasAdded(_ movie: Movie) {
+        movies.append(movie)
+        movieTableView.reloadData()
+        
+    }
+}
+
+
